requirements.txt
Django~=4.1.5
django-phonenumber-field-7.0.2 phonenumbers-8.13.
babel-2.11.0 pytz-2022.7
tinydb-4.7.0
requests==2.28.1
